ENM-3. implemented mechanism for splitting 'name' property
to slowly start shifting away from using 'name' as a way to display location and artist information, two new properties have been created - location and artists - which represent the same information as the current 'name' property but in a clearer and more flexible way. 'name' will either be discontinued or maybe repurposed to contain some other piece of event information. in the meantime, support for both document schemas (documents with 'name' or documents with the additional 'location' and 'artists' properties) will be maintained

additionally, a hardcoded utc timestamp will be given to all events created through EnmEventAdd. this is so newly added events are displayed in the EnmEventListComponent. for testing purposes