DEVOPS-8. renamed some items in .workflow/cicd.yml
the name of the job under this workflow was “build“. in this job were steps not entirely related to building; for example, I saw a step that executed a testing script. instead, since the workflow itself is related to all things CI/CD, the name of the job will be renamed to “cicd“. Should make a little more sense to anyone reading in the future since both building and testing are part of the CI/CD process

also renamed the file so future readers will not be able to compare diffs. overall ok, just a few minor changes