DEVOPS-9. removed some comments from cicd workflow file
just need to retrigger the GitHub Action (which is set to 'on push'). on the initial attempt Vercel was not able to authenticate the GitHub Action workflow in order for a successful deployment to occur. could have been because I wasn't logged into my Vercel account in my local repo. Did a 'vercel login' and now reattempting