ENM-61. added PrimeNG timepicker to EnmEventTimeComponent
had been working on EnmEventDateComponent but realized it can't be finish until its counterpart EnmEventTimeComponent has PrimeNG versions of its input boxes. eventually these two components will merge their data (the date and the time) to create a single timestamp for the database to work with, so i'd rather have them both working off the same PrimeNG model. this commit is a step towards that