ENM-64. added s.t. entire venue object is posted to api and db
added extra typing in the frontend and api EnmEvent models to support a new 'venue' field. this field is where the autocomplete 'Event Location' input box saves. currently the frontend application does not do anything with this field. commit is only interested in having the information saved to the db

note: using 'Event Location' in the UI due to it being more commonly understood than 'Venue'