ENM-75. added temporary mobile-friendly gesture
for improved DX, i should have a quick way of navigating to /add-event/venue. Currently access the route by either typing out the full url in the browser search bar or opening my phone's Notes app which has a hyperlink to the route

for practical reasons i do not want an openly visible 'add event' button quite yet. i do sometimes demo this app to people and in the event they keep access to the web app the route should be somewhat hidden as it still needs to be polished