outsourced container centering logic into directive
wrote some typescript code to center the main container of EnmEventVenueComponent a couple days ago. this makes it s.t. it is in the middle of the screen, horizontally and vertically. code will very likely be reused in subsequent "Add Event" form-related components (i.e. EnmEventDateComponent, EnmEventTimeComponent,  EnmEventPriceOfEntryComponent, EnmEventArtistComponent)

in this commit the code was moved from the venue component to a newly generated directive 'CenterContainerDirective'

i manually removed the test found in the autogenerated spec file for this directive as it was failing (little weird it fails by default)