ENM-72. centered & aligned EnmEventArtistsComponent content/html
not using the centering directive from past commits due to my perceived lack in its flexibility. i need more control over when and why the container becomes centered. This is possible through inlining the centering logic on the component and applying method calls on this logic where needed. i'd have to do more research on directives to find out if it is possible to 'reapply' them on a component, but for now this is a solution

also used ngClass for the first time ever, so there's that. Needed it to dynamically add some passing on some html element