ENM-76. modified s.t. venue suggestions are based off substring
Current behavior is to filter using the string exactly as given by the user. This means searching for a venue “The Swimming Walrus“ with text input “Swimming Walrus” will not yield it as a suggestion. To fix this, an update was made to the filtering function such that it returns suggestions based off substring