modified s.t. keyboard is hidden on EnmEventVenueComponent form submission
the virutal keyboard on mobile devices can sometimes negatively impact UX when navigating between different pages of the event submission form. experimenting with potential solutions. i hope programmatically hiding the sidebar will be a simple fix to the user submitting a subform with the keyboard displayed (which causes misalignments in the subsequent page)