ENM-82. implemented ScrollPanel solution for EnmEventArtistComponent
instead of using PrimeNG's DataView as mentioned in 1b87ecc, a solution using ScrollPanel is used. Turned out too complicated to mess with DataView the way I did. Tried to have it behave sort of like a dynamic Form Array holder to enable the user the ability of adding and removing input fields as necessary. The reason was to avoid having to do any styling myself as most of the time PrimeNG components to a good job of handling that internally. But it got really hacky towards the end

PrimeNG's ScrollPanel provides a good enough solution just by adding a few lines of code in the template. I also get to keep my initial implementation which was already working. Probably should've gone with this approach from the beginning but it was only until I got desperate enough with DataView that I did I hard search through PrimeNG's component list

Should note that ScrollPanel has its problems. When enough inputs are created within the panel and the user scrolls to the top the first input field's label is cut off. It seems to happen conditionally. Only noticed it on Chrome desktop. Tried Safari desktop and worked fine. This commit will let me check mobiles so that will impact the severity of the issue

Still need to remove the packages associated with DataView. Will do that in some later commit