ENM-85. flipped some EEVenueComp conditionals for readability
some parts of this component make distinction between venues selected from the auto complete suggestion list and venues typed by the user that are not part of the list. the conditional currently in place to make this distinction contained slightly more cognitive overhead due to its use of a negative "notFromAutoCompleteSuggestions". instead, it was updated to be inherently positive fromAuto*. should make it easier for devs to read