ENM-91. installed ngrx store and effects packages
thought for a few days about whether to introduce this kind of complexity. there's a lot of mechanisms involved in setting up an ngrx store and im not entirely sure how badly the learning curve is going to stretch me thin

in the end decided on yes. it solves a lot of fairly common problems in web development and although they may not be the most straightforward solutions there's lots of documentation given their widespread use in larger companies and projects