Initial implementation of form value persistence in EnmEventAdd
Super dirty atm but will change soon. At this point in time I got it working for the very first time according to my specifications. Only the first few form input fields are being stored in NgRx but as time passes and as I develop a better implementation I'll add the rest of them. The input fields are not selecting on this NgRx state either. So for now, it really is just bare bone functionality - just sure state is being saved correctly and that's it