ENM-122. reenabled event detail page
It's a rough implementation for now. Though I do want this to be a plus feature so I made relevant changes making it is only accessible for these users

Something worth noting here is I had difficulty getting the list to render in the case where no user is logged in. Trouble stemmed from how null values are handled by the, I think, *ngIf operator. When the value its condition is based on returns null, then the entire element does not render. There's probably a cleaner workaround, but I opted for a copy/paste-esque solution for now. Maybe in the future can look into the details more