ENM-125. Meta reducer hydration imp complete
Surprised at how easy that was. Got it right the first time w no bugs. Though there is some weird behavior and I need to take a look at. One is that the filter, upon refresh, maintains the text it was using. Meaning although the refresh wipes the visual text that goes in the search box for the user to see, the app maintains the text in the background and continues to filter by that text. Another thing I need to take a look at is the form. It might be that I don't want that being stored in state. Might benefit just removing that from localStorage completely

Now that I have this working the next step is to make sure when on the detail page and when the user refreshes that the detail event page does not go blank (i.e., the component should probably be listening to the store for the selected event and continuously display that)