ENM-125. EEPageComp st it listens to NgRx store for event IDs
Long ways to get here, but okay now users can refresh on the detail page and they will still be able to see the event (before the screen would blank out). As a side effect, most of the event listing stuff (including the detail page) is now based on an NgRx/localStorage implementation. That is to say the application has better state managing than before ENM-125 was implemented. Before this I think the app only persisted session related stuff (username, plus status, etc)

Some things to note here:
- The hydration implementation is completely derived from: https://github.com/nilsmehlhorn/ngrx-issue-tracker/tree/791f608ef69a3a5e1417a3677cc2ba1c05d0fbbc
It's a rather simple and crude implementation of hydration but it'll do for now. Also I got a better grasp of meta reducers in the process of implementing this so that's v good
- I tried a custom implementation that involved tapping into store directly from app component and persisting/rehydrating that way. I didn't get very far as it started biting back really quickly. I spent most of yesterday trying to get it working but bleh. They key to the implementation I ended up going with is that it uses actions and the effects lifecycle hooks which gives a lot more flexibility as to when state gets set