reimplemented EEListComp st its using PrimeNG's DataView
Kind of a big change. Expecting at least *something* to break after pushing. Doing this mainly because if I use PrimeNG components for this page they'll by default use the styling I've applied globally for the app. Which is nice. The colors before looked real out of place compared to other components on the page like the toolbar and sidebar button. Hopefully this makes it easier on the eyes