ENM-124. Reimplemented artist property on EnmEvent model
Big change sorta. Up until now the artists property on EnmEvent model had been a simple string array where strings represented artist names. I'm interested in providing links to these artists social media pages, so I think a good step forward is to add a little complexity on this array by creating a dedicated 'Artist' type that will hold not only their name but also a property which links to their instagrams. In the future this might be beneficial if I want to extend the functionality of this artist type further

Some things to note. At the moment there is no way to create these artist objects through the application (in other words artists can only be choses from the predefined autocomplete suggestions provided by the application). I'd need to go in the db and manually create an artist every time I come across an artist that has yet to be implemented in the db. Not too big of a deal in all honesty since at some point in the future I'd have gotten just about every artist in the valley, though a mechanism to do this automatically would be nice to have. Another thing to note is that the app breaks the api if an artist is submitted that is not listed in the autocomplete suggestions. I'd also need to take a closer look at this, but given that I will be the only person using this form for the distant future I think it'll be fine for now. Just have to remember to always click on a suggestion

In general I found this change a little difficult. Not because it was technically difficult but because of the design choices I had to make. For example, it's clearer to me now that mongodb (and I guess document dbs in general) have the effect of forcing your application to deal with schema changes by continuously adding template code that not only handles the new schema but is also backwards compatible with the older versions. If you don't plan on updating your database documents, that is. Kind of a difficult choice since I don't know what the long term affects are. In the end I decided to take on a cut throat attitude towards the documents in mongodb. Basically "idc if I lose backward compatibility. Press forward and forward only". Probably in the future I'll spend some time cleaning the older documents so that they conform to the most recent schema. In the distant future I kinda think it would be cool to offer some kind of 'history' feature where users can go back and look at all the events that have happened throughout the years so I want to make sure the application can process all events in the collection