Explicit lazy loading on AppRoutingModule
Hm, not entirely sure if this is doing something. But without this line it doesn't make sense to me how the app is able to figure out how to route to the listing component (EEListComp). Maybe because the component is associated to a blank route ('') and so it defaults to it? Idk it's confusing since I thought somehow AppRoutingModule needed a reference to EnmEventRoutingModule for everything to come together. Either way, explicitly writing down the lazy loading of EnmEvents helps with my mental model of the app's routing logic