Restructured routing paths
Enm event related paths were their own routing module. Consolidating it now with app routing module. Doing this because it might help with preserving scroller state when navigating between the event list and the detail page of an event. Trying to follow this example from SO but having some trouble due to the lazy loaded nature of the enm routing module: https://stackoverflow.com/questions/41280471/how-to-implement-routereusestrategy-shoulddetach-for-specific-routes-in-angular