implemented RouteReuseStrategy on EEListComp
The end goal is extremely fluid navigation between list and detail page (and vice versa). To this end, we want to make sure the app reutilizes EEListComp instead of creating a new instance (along with associated networks calls). This is a test run. Just want to see how it behaves on mobile