List http request sent after logging in
Issue where logging in wasn't triggering the HTTP request in charge of getting the list of events. Problematic when plus users logged out then logged back in. Doing this sequence of actions results in plus users receiving the non-plus version of the list

This fix has a tiny drawback in that in makes two calls instead of one. For now that's okay I'll have to take a look at it in the future to investigate further. But at least plus users won't have issues viewing the entire event listing