ENM-178. Sidebar on event page too
Instead of just on the list.

Note that I just copy/pasted code but my intent for long term is to somehow structure HTML s.t. there is only one instance of the toolbar code which is shared by both components. Just thought this would work in a pinch since launch date is coming relatively soon